Restructure move feature classes to tools package
